mejorado

+++ src/main/java/com/fisisac/developer/fisisackback/repository/ProgramaRepository.java
@@ -3,6 +3,6 @@ package com.fisisac.developer.fisisackback.repository;
 import com.fisisac.developer.fisisackback.domain.Programa;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProgramaRepository extends CrudRepository<Programa, Long> {
+public interface ProgramaRepository extends CrudRepository<Programa, Short> {
 
 }
